Instead of having a bunch of bundles, we now have 2 bundles. A full bundle, called app.bundle.js and a smaller bundle called takeReader.bundle.js.
The takeReader bundle contains just enough for a user reading someone else's Take for the first time. It does not include the Timeline component, which is the largest component at almost 2 MB by itself. This improves the experience for new MyTake users because it decreases initial load time.

All other routes load the full bundle, app.bundle.js.

Also renamed documentReader.tsx to takeReader.tsx in this commit.
