+++ nunjucks/pages/foundation/bill-of-rights/index.html
@@ -1,6 +1,5 @@
-{% extends "foundationExplorer.html" %}
+{% extends "react.html" %}
 
 {% block content %}
-  <link href="/vis.css" rel="stylesheet" type="text/css" />
   <div id="app"></div>
 {% endblock %}
