+++ nunjucks/pages/index.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "takeReader.html" %}
 
 {% block content %}
 <div class="hero__wrapper">
