+++ nunjucks/pages/new-take/index.html
@@ -1,6 +1,5 @@
 {% extends "react.html" %}
 
 {% block content %}
-  <link href="/vis.css" rel="stylesheet" type="text/css" />
   <div id="app"></div>
 {% endblock %}
