+++ nunjucks/pages/samples/does-a-law-mean-what-it-says-or-what-it-meant/index.html
@@ -1,4 +1,4 @@
-{% extends "documentReader.html" %}
+{% extends "takeReader.html" %}
 
 {% block content %}
   <div id="app"></div>
