+++ nunjucks/pages/samples/dont-worry-well-protect-the-constitution-for-you/index.html
@@ -1,4 +1,4 @@
-{% extends "documentReader.html" %}
+{% extends "takeReader.html" %}
 
 {% block content %}
   <div id="app"></div>
