+++ nunjucks/templates/blockEditor.html
@@ -2,5 +2,5 @@
 
 {% block bundle %}
   <!-- Main -->
-  <script src="/blockEditor.bundle.js"></script>
+  <script src="/blockEditorTest.bundle.js"></script>
 {% endblock %}
