--- nunjucks/templates/documentReader.html
@@ -1,6 +1,0 @@
-{% extends "reactCore.html" %}
-
-{% block bundle %}
-  <!-- Main -->
-  <script src="/documentReader.bundle.js"></script>
-{% endblock %}
