--- nunjucks/templates/foundation.html
@@ -1,6 +1,0 @@
-{% extends "reactCore.html" %}
-
-{% block bundle %}
-  <!-- Main -->
-  <script src="/foundation.bundle.js"></script>
-{% endblock %}
