+++ nunjucks/templates/react.html
@@ -2,5 +2,6 @@
 
 {% block bundle %}
 	<!-- Main -->
+	<link href="/vis.css" rel="stylesheet" type="text/css" />
 	<script src="/app.bundle.js"></script>
 {% endblock %}
