+++ nunjucks/templates/takeReader.html
@@ -2,5 +2,5 @@
 
 {% block bundle %}
   <!-- Main -->
-  <script src="/foundationExplorer.bundle.js"></script>
+  <script src="/takeReader.bundle.js"></script>
 {% endblock %}
