+++ src/takeReader.tsx
